backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.29.0
bcrypt==5.0.0
beautifulsoup4==4.12.2
better-auth==0.0.1b11
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
distro==1.9.0
fastapi==0.104.1
greenlet==3.3.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
hyperframe==6.1.0
idna==3.11
Markdown==3.5.1
numpy==2.3.5
openai==1.3.6
portalocker==2.10.1
protobuf==6.33.2
psycopg2-binary==2.9.9
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
qdrant-client==1.8.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.23
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
